+++ main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from services.study import router as ConfigRouter
+from routers.study_config_routes import router as ConfigRouter
 import uvicorn as uv
 
 #Initialize FastAPI App
@@ -9,7 +9,9 @@ app.include_router(ConfigRouter)
 
 @app.get("/")
 def hello_world():
-    return "Hello World"
+    return {"message": "Hello World"}
+
 
 if __name__ == "__main__":
     uv.run(app, host="0.0.0.0")
+
