+++ models/base.py
@@ -1,0 +1,2 @@
+from sqlalchemy.orm import declarative_base
+Base = declarative_base()
