+++ models/enums.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class DisplayMethodEnum(str,Enum):
+    SEQUENTIAL = "sequential"
+    RANDOM = "random"
+
+class ResponseMethodEnum(str,Enum):
+    BINARY = "binary"
+    GRADIENT = "gradient"
